Added new files and fixed errors
