+++ src/modules/constants.js
@@ -1,0 +1,8 @@
+const todoInput = document.querySelector('#newtodo');
+const todosListEl = document.querySelector('#todos-list');
+const form = document.querySelector('#todoform');
+const todo = 0;
+
+export {
+  todoInput, todosListEl, form, todo,
+};
