+++ src/template.html
@@ -9,7 +9,7 @@
       rel="stylesheet"
       href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.8.3/font/bootstrap-icons.css"
     />
-    <!-- <link rel="stylesheet" href="./style.css" /> -->
+    <link rel="stylesheet" href="./mai.css" />
   </head>
   <body>
     <div class="container">
@@ -46,6 +46,6 @@
       <div class="notification"></div>
     </div>
 
-    <!-- <script src="./script.js"></script> -->
+    <script type="module" src="./index.js"></script>
   </body>
 </html>
